Moved run function to blackwidow.py from sim.py
