+++ src/blackwidow/__init__.py
@@ -1,1 +1,1 @@
-from sim import run
+from blackwidow import run
