+++ src/blackwidow/sim/__init__.py
@@ -1,1 +1,1 @@
-from sim import run
+
